+++ awscli/errorhandler.py
@@ -10,8 +10,6 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-
-import sys
 import logging
 
 LOG = logging.getLogger(__name__)
@@ -54,11 +52,8 @@ class ErrorHandler(object):
     with an appropriate error code.
     """
 
-    def __call__(self, http_response, parsed, operation, **kwargs):
+    def __call__(self, http_response, parsed, model, **kwargs):
         LOG.debug('HTTP Response Code: %d', http_response.status_code)
-        msg_template = ("A {error_type} error ({error_code}) occurred "
-                        "when calling the {operation_name} operation: "
-                        "{error_message}")
         error_type = None
         error_class = None
         if http_response.status_code >= 500:
@@ -71,7 +66,7 @@ class ErrorHandler(object):
             code, message = self._get_error_code_and_message(parsed)
             raise error_class(
                 error_code=code, error_message=message,
-                error_type=error_type, operation_name=operation.name,
+                error_type=error_type, operation_name=model.name,
                 http_status_code=http_response.status_code)
 
     def _get_error_code_and_message(self, response):